refactor: portfolio__link-arrow нужно вложить в ссылку
