+++ src/components/AboutMe/AboutMe.css
@@ -9,7 +9,7 @@
   letter-spacing: -0.88px;
   border-bottom: 1px solid rgba(0, 0, 0, 1);
   margin: 0 70px;
-  padding: 111px 0 23px 0;
+  padding: 113px 0 23px 0;
 }
 
 @media screen and (max-width: 1279px) {
