+++ src/components/AboutProject/AboutProject.css
@@ -52,7 +52,7 @@
   .about-project__info {
     flex-direction: column;
     margin: 0 18px;
-    padding: 61px 0 0;
+    padding: 62px 0 0;
   }
 }
