+++ src/components/Portfolio/Portfolio.css
@@ -46,7 +46,7 @@
 
 @media screen and (max-width: 767px) {
   .portfolio__list {
-    margin: 40px 0 70px;
+    margin: 41px 0 70px;
     padding: 0 14px;
   }
 }
@@ -61,22 +61,29 @@
   padding: 0;
 }
 
-.portfolio__list-item:first-child {
-  margin: 0;
-}
-
-.portfolio__list-item:last-child {
-  border-bottom: none;
+@media screen and (max-width: 1279px) {
+  .portfolio__list-item {
+    margin: 20px 0 0;
+  }
 }
 
 @media screen and (max-width: 767px) {
   .portfolio__list-item {
+    margin: 21px 0 0;
     padding: 0;
   }
 
   .portfolio__list-item:last-child {
     padding: 0;
   }
+}
+
+.portfolio__list-item:first-child {
+  margin: 0;
+}
+
+.portfolio__list-item:last-child {
+  border-bottom: none;
 }
 
 .portfolio__link {
@@ -86,14 +93,19 @@
   font-weight: 400;
   line-height: 50px;
   letter-spacing: -1.2px;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  width: 100%;
   margin: 0;
-  padding: 0 0 19px;
+  padding: 0 0 8px;
 }
 
 @media screen and (max-width: 1279px) {
   .portfolio__link {
     font-size: 28px;
     letter-spacing: -1.12px;
+    padding: 0 0 8px;
   }
 }
 
@@ -102,6 +114,7 @@
     font-size: 18px;
     line-height: 28px;
     letter-spacing: -0.72px;
+    padding: 0 0 18px;
   }
 }
