+++ src/components/Techs/Techs.css
@@ -33,7 +33,7 @@
     font-size: 18px;
     letter-spacing: -0.72px;
     margin: 0 18px;
-    padding: 37px 0 28px 0;
+    padding: 36px 0 28px 0;
   }
 }
